Move `intersection::surface` to dedicated module
